requirements.txt
actdiag
alembic>=0.6.4
blockdiag
docutils==0.11
nwdiag
oslosphinx
pecan>=0.7.0
pbr>=0.6,<1.0
seqdiag
sphinx
sphinxcontrib-actdiag
sphinxcontrib-blockdiag
sphinxcontrib-nwdiag
sphinxcontrib-seqdiag
SQLAlchemy>=0.7.8,<=0.9.99
graphviz

anyjson>=0.3.3
Babel>=1.3
eventlet>=0.13.0
requests>=1.2.1
jsonrpclib
netaddr>=0.7.6
python-neutronclient>=2.3.6,<3
WebOb>=1.2.3
six>=1.7.0
oslo.config>=1.4.0.0a3
oslo.db>=0.4.0  # Apache-2.0
oslo.messaging>=1.4.0.0a3
oslo.rootwrap>=1.3.0.0a1
python-novaclient>=2.17.0
posix_ipc
wsme==0.6.1
